example-python/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.1.0
langchain-core>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0
